requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
Flask==3.0.1
fonttools==4.44.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
proglog==0.1.10
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
urllib3==2.2.0
Werkzeug==3.0.1
